Improve dynamic voice creation & deletion, update discord.js
Divide 'ready' event in sub-functions, fix issue preventing the deletion of a channel when its last user switches from a channel to another instead of disconnecting, sleep for 2.5s to prevent spam, load channels beginning with specific prefix in case of crash to continue handling channel deletion ; update discord.js to v12.3.1

+++ events/voiceChannelJoin.js
@@ -1,31 +1,30 @@
-const { dynamicVoiceChannelids } = require("../config.js");
+const { dynamicVoiceChannelPrefix, dynamicVoiceChannelids } = require('../config.js');
 
 module.exports = async (client, log, member, channel) => {
-    if (dynamicVoiceChannelids.includes(channel.id)) {
-        log.info("Member wants to create a dynamic voice channel");
-        let newChannel;
+  if (dynamicVoiceChannelids.includes(channel.id)) {
+    await new Promise((resolve) => setTimeout(resolve, 2500)); // sleep to prevent overuse
 
-        // if the user leaves before the new channel was created, moving him to the new channel will crash
-        // try catch to delete the channel if it exists but the creation failed
-        try {
-            const newOptions = {
-                name: `Salle de ${member.nickname}`,
-                parent: channel.parentID,
-                permissionOverwrites: [{
-                  id: member.id, 
-                  allow: ["MANAGE_CHANNELS"],
-                  type: "member"
-                }]
-              };
-            newChannel = await channel.clone(newOptions);
-            client.dynamicVoiceChannels.set(newChannel.id, newChannel);
-            await member.voice.setChannel(newChannel);
-        } catch (e) {
-            log.error("There was an error creating a dynamic voice channel  : " + e);
-            if (newChannel) {
-                newChannel.delete();
-            }
-        }
-        
+    log.info(`Member <${member.nickname}> (${member.id}) wants to create a dynamic voice channel`);
+    
+    let newChannel;
+    try {
+      const newOptions = {
+        name: `${dynamicVoiceChannelPrefix} ${member.nickname}`,
+        parent: channel.parentID,
+        permissionOverwrites: [{
+          id: member.id,
+          allow: ['MANAGE_CHANNELS'],
+          type: 'member'
+        }]
+      };
+      newChannel = await channel.clone(newOptions);
+      client.dynamicVoiceChannels.set(newChannel.id, newChannel);
+
+      await member.voice.setChannel(newChannel);
+    } catch (e) {
+      // if the user leaves before the new channel was created, moving him to the new channel will crash
+      log.error('There was an error creating a dynamic voice channel: ' + e);
+      if (newChannel) newChannel.delete();
     }
+  }
 };
